@@ -20,12 +20,13 @@
     "build.demo": "ng build --prod",
     "build.package": "ng-packagr -p ng-package.json",
     "lint": "ng lint",
+    "lint:fix": "ng lint --fix",
     "ng": "ng",
     "precommit": "npm run test",
     "release": "standard-version --commit-all",
     "publish.ng4": "npm publish dist --tag ng4",
-    "publish.ng9": "npm publish dist",
-    "publish.ng9-dev": "npm publish dist --tag dev",
+    "publish.ng10": "npm publish dist",
+    "publish.ng10-dev": "npm publish dist --tag dev",
     "start": "ng serve",
     "test": "ng test --watch false",
     "test-coverage": "ng test --watch false --code-coverage"
@@ -51,48 +52,49 @@
   ],
   "peerDependencies": {},
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.1",
-    "@angular/animations": "~9.1.1",
-    "@angular/cli": "~9.1.1",
-    "@angular/common": "~9.1.1",
-    "@angular/compiler": "~9.1.1",
-    "@angular/compiler-cli": "~9.1.1",
-    "@angular/core": "~9.1.1",
-    "@angular/forms": "~9.1.1",
-    "@angular/language-service": "~9.1.1",
-    "@angular/platform-browser": "~9.1.1",
-    "@angular/platform-browser-dynamic": "~9.1.1",
-    "@angular/router": "~9.1.1",
-    "@types/jasmine": "~3.3.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/marked": "^0.3.0",
-    "@types/node": "^12.11.1",
-    "@types/webpack": "^4.4.34",
-    "codecov": "^3.0.0",
-    "codelyzer": "^5.1.2",
-    "core-js": "^2.5.4",
-    "escape-string-regexp": "^1.0.5",
-    "html-loader": "^0.5.5",
-    "husky": "^0.14.3",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
+    "@angular-devkit/build-angular": "~0.1002.0",
+    "@angular/animations": "~10.2.3",
+    "@angular/cli": "~10.2.0",
+    "@angular/common": "~10.2.3",
+    "@angular/compiler": "~10.2.3",
+    "@angular/compiler-cli": "~10.2.3",
+    "@angular/core": "~10.2.3",
+    "@angular/forms": "~10.2.3",
+    "@angular/language-service": "~10.2.3",
+    "@angular/platform-browser": "~10.2.3",
+    "@angular/platform-browser-dynamic": "~10.2.3",
+    "@angular/router": "~10.2.3",
+    "@types/jasmine": "^3.6.1",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/marked": "^1.1.0",
+    "@types/node": "^14.14.7",
+    "@types/webpack": "^4.41.24",
+    "codecov": "^3.8.1",
+    "codelyzer": "^6.0.1",
+    "escape-string-regexp": "^4.0.0",
+    "html-loader": "^1.3.2",
+    "husky": "^4.3.0",
+    "jasmine-core": "^3.6.0",
+    "jasmine-spec-reporter": "^6.0.0",
+    "karma": "^5.2.3",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "lodash.isequal": "^4.5.0",
-    "markdown-loader": "^5.1.0",
-    "ng-packagr": "^9.0.0",
-    "ngx-bootstrap": "^5.6.1",
+    "markdown-loader": "^6.0.0",
+    "ng-packagr": "^10.1.0",
+    "ngx-bootstrap": "^6.1.0",
     "pre-commit": "^1.2.2",
-    "protractor": "~5.4.0",
-    "rxjs": "~6.5.5",
-    "standard-version": "^7.0.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.8.3",
-    "zone.js": "~0.10.2"
+    "protractor": "~7.0.0",
+    "rxjs": "~6.6.3",
+    "standard-version": "^9.0.0",
+    "ts-node": "^9.0.0",
+    "tslib": "^2.0.3",
+    "tslint": "~6.1.0",
+    "tslint-jasmine-rules": "^1.6.1",
+    "typescript": "~4.0.5",
+    "zone.js": "^0.11.3"
   },
   "contributors": [
     {
@@ -115,8 +117,5 @@
       "email": "otelnov@gmail.com",
       "url": "https://github.com/otelnov"
     }
-  ],
-  "dependencies": {
-    "tslib": "^1.10.0"
-  }
+  ]
 }
