@@ -1,21 +1,21 @@
-import {TestBed, async, fakeAsync, tick} from '@angular/core/testing';
-import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-import {CommonModule} from '@angular/common';
-import {BrowserModule} from '@angular/platform-browser';
-import {AppComponent} from './app.component';
-import {NgxSelectModule} from './lib/public_api';
-import {SelectSectionComponent} from './demo/select-section';
-import {SampleSectionComponent} from './demo/sample-section.component';
-import {SingleDemoComponent} from './demo/select/single-demo';
-import {RichDemoComponent} from './demo/select/rich-demo';
-import {MultipleDemoComponent} from './demo/select/multiple-demo';
-import {ChildrenDemoComponent} from './demo/select/children-demo';
-import {NoAutoCompleteDemoComponent} from './demo/select/no-autocomplete-demo';
-import {TabsModule} from 'ngx-bootstrap/tabs';
-import {ButtonsModule} from 'ngx-bootstrap/buttons';
+import { TestBed, fakeAsync, tick, waitForAsync } from '@angular/core/testing';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { CommonModule } from '@angular/common';
+import { BrowserModule } from '@angular/platform-browser';
+import { AppComponent } from './app.component';
+import { NgxSelectModule } from './lib/public_api';
+import { SelectSectionComponent } from './demo/select-section';
+import { SampleSectionComponent } from './demo/sample-section.component';
+import { SingleDemoComponent } from './demo/select/single-demo';
+import { RichDemoComponent } from './demo/select/rich-demo';
+import { MultipleDemoComponent } from './demo/select/multiple-demo';
+import { ChildrenDemoComponent } from './demo/select/children-demo';
+import { NoAutoCompleteDemoComponent } from './demo/select/no-autocomplete-demo';
+import { TabsModule } from 'ngx-bootstrap/tabs';
+import { ButtonsModule } from 'ngx-bootstrap/buttons';
 
 describe('AppComponent', () => {
-    beforeEach(async(() => {
+    beforeEach(waitForAsync(() => {
         TestBed.configureTestingModule({
             imports: [
                 BrowserModule,
@@ -24,7 +24,7 @@ describe('AppComponent', () => {
                 NgxSelectModule,
                 TabsModule.forRoot(),
                 ButtonsModule.forRoot(),
-                CommonModule
+                CommonModule,
             ],
             declarations: [
                 AppComponent,
@@ -34,7 +34,7 @@ describe('AppComponent', () => {
                 MultipleDemoComponent,
                 NoAutoCompleteDemoComponent,
                 RichDemoComponent,
-                SingleDemoComponent
+                SingleDemoComponent,
             ],
         }).compileComponents();
     }));
