@@ -1,9 +1,9 @@
-import {Component, Input} from '@angular/core';
+import { Component, Input } from '@angular/core';
 
 @Component({
     selector: 'sample-section',
-    templateUrl: './sample-section.component.html'
+    templateUrl: './sample-section.component.html',
 })
 export class SampleSectionComponent {
-    @Input() public desc: any;
+    @Input() public desc: { heading: string, html: { default: string }, ts: { default: string } };
 }
